+++ projects/packages/stats/src/class-package-version.php
@@ -12,7 +12,7 @@ namespace Automattic\Jetpack\Stats;
  */
 class Package_Version {
 
-	const PACKAGE_VERSION = '0.15.1';
+	const PACKAGE_VERSION = '0.15.3';
 
 	const PACKAGE_SLUG = 'stats';