+++ projects/packages/status/src/class-host.php
@@ -147,6 +147,10 @@ class Host {
 		}
 
 		$dns_records = dns_get_record( $domain, DNS_NS ); // Fetches the DNS records of type NS (Name Server)
+		if ( false === $dns_records ) {
+			return array();
+		}
+
 		$nameservers = array();
 		foreach ( $dns_records as $record ) {
 			if ( isset( $record['target'] ) ) {
@@ -235,7 +239,7 @@ class Host {
 	/**
 	 * Returns a guess of the hosting provider for the current site based on various checks.
 	 *
-	 * @since $$next-version$$ Added $guess parameter.
+	 * @since 5.0.4 Added $guess parameter.
 	 *
 	 * @param bool $guess Whether to guess the hosting provider.
 	 *