+++ projects/plugins/jetpack/class-jetpack-newsletter-dashboard-widget.php
@@ -5,23 +5,18 @@
  * @package jetpack
  */
 
-use Automattic\Jetpack\Assets;
-
 /**
  * Adds the Jetpack Newsletter widget to the WordPress admin dashboard.
  *
  * @package jetpack
  */
 
+use Automattic\Jetpack\Connection\Client;
+
 /**
  * Class that adds the Jetpack Newsletter Dashboard Widget to the WordPress admin dashboard.
  */
 class Jetpack_Newsletter_Dashboard_Widget {
-	const JS_DEPENDENCIES = array( 'lodash', 'react', 'react-dom', 'wp-api-fetch', 'wp-components', 'wp-compose', 'wp-element', 'wp-html-entities', 'wp-i18n', 'wp-is-shallow-equal', 'wp-polyfill', 'wp-primitives', 'wp-url', 'wp-warning', 'moment' );
-	// Sometimes custom scripts would strip the `ver` query params, so we need to make sure it doesn't by adding a custom version param `osv` here.
-	const NEWSLETTER_WIDGET_CDN_URL = 'https://widgets.wp.com/newsletter/%s?minify=false';
-	const NEWSLETTER_WIDGET_VERSION = '1.0.0';
-
 	/**
 	 * Indicates whether the class initialized or not.
 	 *
@@ -49,11 +44,54 @@ class Jetpack_Newsletter_Dashboard_Widget {
 	}
 
 	/**
+	 * Get the config data for the Jetpack Newsletter widget.
+	 *
+	 * @return array
+	 */
+	public static function get_config_data() {
+		$subscriber_counts = array();
+		$config_data       = array();
+
+		if ( Jetpack::is_connection_ready() ) {
+			$site_id  = Jetpack_Options::get_option( 'id' );
+			$api_path = sprintf( '/sites/%d/subscribers/counts', $site_id );
+			$response = Client::wpcom_json_api_request_as_blog(
+				$api_path,
+				'2',
+				array(),
+				null,
+				'wpcom'
+			);
+
+			if ( 200 === wp_remote_retrieve_response_code( $response ) ) {
+				$subscriber_counts = json_decode( wp_remote_retrieve_body( $response ), true );
+				if ( isset( $subscriber_counts['counts']['email_subscribers'] ) ) {
+					$config_data['emailSubscribers'] = (int) $subscriber_counts['counts']['email_subscribers'];
+				}
+
+				if ( isset( $subscriber_counts['counts']['paid_subscribers'] ) ) {
+					$config_data['paidSubscribers'] = (int) $subscriber_counts['counts']['paid_subscribers'];
+				}
+			}
+		}
+
+		return $config_data;
+	}
+
+	/**
 	 * Sets up the Jetpack Newsletter widget in the WordPress admin dashboard.
 	 */
 	public static function wp_dashboard_setup() {
-		static::load_admin_scripts( 'jp-newsletter-widget', 'newsletter.min', array( 'config_variable_name' => 'jetpackNewsletterWidgetConfigData' ) );
 		if ( Jetpack::is_connection_ready() ) {
+			static::load_admin_scripts(
+				'jp-newsletter-widget',
+				'newsletter-widget',
+				array(
+					'config_variable_name' => 'jetpackNewsletterWidgetConfigData',
+					'config_data'          => static::get_config_data(),
+				)
+			);
+
 			$widget_title = sprintf(
 				__( 'Newsletter', 'jetpack' )
 			);
@@ -73,7 +111,7 @@ class Jetpack_Newsletter_Dashboard_Widget {
 	 */
 	public static function render() {
 		?>
-		<div id="wpcom" style="min-height: calc(100vh - 100px);">
+		<div id="wpcom">
 			<div id="newsletter-widget-app"></div>
 		</div>
 		<?php
@@ -85,7 +123,14 @@ class Jetpack_Newsletter_Dashboard_Widget {
 	 * @return void
 	 */
 	public static function admin_init() {
-		static::load_admin_scripts( 'jp-newsletter-widget', 'newsletter.min', array( 'config_variable_name' => 'jetpackNewsletterWidgetConfigData' ) );
+		static::load_admin_scripts(
+			'jp-newsletter-widget',
+			'newsletter-widget',
+			array(
+				'config_variable_name' => 'jetpackNewsletterWidgetConfigData',
+				'config_data'          => static::get_config_data(),
+			)
+		);
 	}
 
 	/**
@@ -103,30 +148,57 @@ class Jetpack_Newsletter_Dashboard_Widget {
 			'enqueue_css'          => true,
 		);
 		$options         = wp_parse_args( $options, $default_options );
-		if ( file_exists( __DIR__ . "/../dist/{$asset_name}.js" ) ) {
-			// Load local assets for the convinience of development.
-			Assets::register_script(
-				$asset_handle,
-				"../dist/{$asset_name}.js",
-				__FILE__,
-				array(
-					'in_footer'  => true,
-					'textdomain' => 'jetpack',
-				)
-			);
-			Assets::enqueue_script( $asset_handle );
-		} else {
-			// In production, we load the assets from our CDN.
-			wp_register_script(
-				$asset_handle,
-				sprintf( self::NEWSLETTER_WIDGET_CDN_URL, "{$asset_name}.js" ),
-				self::JS_DEPENDENCIES,
-				self::NEWSLETTER_WIDGET_VERSION,
-				true
-			);
-			wp_enqueue_script( $asset_handle );
+
+		// Get the asset file path
+		$asset_path = JETPACK__PLUGIN_DIR . '_inc/build/' . $asset_name . '.min.asset.php';
+
+		// Get dependencies and version from asset file
+		$dependencies = array();
+		$version      = JETPACK__VERSION;
+
+		if ( file_exists( $asset_path ) ) {
+			$asset        = require $asset_path;
+			$dependencies = $asset['dependencies'];
+			$version      = $asset['version'];
 		}
 
-		// TODO: Add config data.
+		// Register and enqueue the script
+		wp_register_script(
+			$asset_handle,
+			plugins_url( '_inc/build/' . $asset_name . '.min.js', JETPACK__PLUGIN_FILE ),
+			$dependencies,
+			$version,
+			true
+		);
+		wp_enqueue_script( $asset_handle );
+
+		// Enqueue the CSS if enabled
+		if ( $options['enqueue_css'] ) {
+			wp_enqueue_style(
+				$asset_handle,
+				plugins_url( '_inc/build/' . $asset_name . '.css', JETPACK__PLUGIN_FILE ),
+				array(),
+				$version
+			);
+
+			// Enqueue RTL stylesheet if needed
+			if ( is_rtl() ) {
+				wp_enqueue_style(
+					$asset_handle . '-rtl',
+					plugins_url( '_inc/build/' . $asset_name . '.rtl.css', JETPACK__PLUGIN_FILE ),
+					array( $asset_handle ),
+					$version
+				);
+			}
+		}
+
+		// Add any configuration data if needed
+		if ( ! empty( $options['config_data'] ) ) {
+			wp_add_inline_script(
+				$asset_handle,
+				"window.{$options['config_variable_name']} = " . wp_json_encode( $options['config_data'] ) . ';',
+				'before'
+			);
+		}
 	}
 }